Fixed issue with 16 bit ints on Arduino where FindGlyph failed to locate glyphs in the last CharcodeRun

+++ Arduino/libraries/XFont/XFont.cpp
@@ -152,7 +152,7 @@ uint16_t XFont::FindGlyph(
 		{
 			current = (leftIndex + rightIndex) / 2;
 			
-			int	cmpResult = pgm_read_word_near(&charcodeRuns[current].start) - inCharcode;
+			int32_t	cmpResult = (int32_t)pgm_read_word_near(&charcodeRuns[current].start) - inCharcode;
 			if (cmpResult == 0)
 			{
 				charcodeRun = &charcodeRuns[current];
